Task 3
Created a lower and upper bound to track the correlation between two stocks.
We track the price of each stock at a given time based on the array supplied by the server.
Also set a trigger alert that will go off if the ratio between the two stocks are +/- 10% their historical movements.